Move lotus json and rpc context under utils
