+++ src/faucet/calibnet/views.rs
@@ -2,8 +2,8 @@ use crate::faucet::calibnet::{
     CALIBNET_DRIP_AMOUNT, CALIBNET_RATE_LIMIT_SECONDS, FIL_CALIBNET_UNIT,
 };
 use crate::faucet::views::faucet::Faucet;
-use crate::rpc_context::{Provider, RpcContext};
 use crate::utils::format::format_balance;
+use crate::utils::rpc_context::{Provider, RpcContext};
 use fvm_shared::address::Network;
 use leptos::prelude::*;
 use leptos::{component, view, IntoView};
