+++ src/utils/lotus_json/mod.rs
@@ -186,7 +186,7 @@ mod raw_bytes; // fvm_ipld_encoding::RawBytes: !quickcheck::Arbitrary
 #[serde(rename_all = "PascalCase")]
 pub struct MessageLookup {
     pub height: i64,
-    #[serde(with = "crate::lotus_json")]
+    #[serde(with = "crate::utils::lotus_json")]
     pub message: Cid,
 }
 lotus_json_with_self!(MessageLookup);
@@ -362,7 +362,7 @@ impl<T> LotusJson<T> {
 macro_rules! lotus_json_with_self {
     ($($domain_ty:ty),* $(,)?) => {
         $(
-            impl $crate::lotus_json::HasLotusJson for $domain_ty {
+            impl $crate::utils::lotus_json::HasLotusJson for $domain_ty {
                 type LotusJson = Self;
                 fn into_lotus_json(self) -> Self::LotusJson {
                     self
