+++ src/utils/mod.rs
@@ -2,4 +2,6 @@ pub mod address;
 pub mod error;
 pub mod format;
 pub mod key;
+pub mod lotus_json;
 pub mod message;
+pub mod rpc_context;
